requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.9
pydantic==2.7.4
